# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "博邦 ERP Echo + GORM 模块化单体后台接口文档。后续新增、修改、删除接口时，必须同步更新 Swagger 注释、docs/API.md 和 test.http。",
        "title": "博邦 ERP 管理系统 API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/products": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "查询产品",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/bb_erp_echo_internal_model.Product"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_product.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_product.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_product.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/internal_product.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_product.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建产品主数据；必须选择当前账号部门下的在职员工。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "创建产品",
                "parameters": [
                    {
                        "description": "产品参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_product.createProductRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/bb_erp_echo_internal_model.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_product.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_product.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_product.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/internal_product.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_product.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/workorder/products": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建启用状态的正式产品档案；初始安全库存和当前库存均为 0，不创建库存流水。接口同时需要 workorder:write 和 workorder:temporary-product:write 权限。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workorder"
                ],
                "summary": "临时建立仓库产品档案",
                "parameters": [
                    {
                        "description": "产品建档参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_workorder.temporaryProductRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/bb_erp_echo_internal_model.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_workorder.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_workorder.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_workorder.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_workorder.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/internal_workorder.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "bb_erp_echo_internal_model.Product": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "default_cost": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "operator_department_id": {
                    "type": "integer"
                },
                "operator_department_name": {
                    "type": "string"
                },
                "operator_employee_id": {
                    "type": "integer"
                },
                "operator_employee_name": {
                    "type": "string"
                },
                "operator_terminal_id": {
                    "type": "integer"
                },
                "operator_user_id": {
                    "type": "integer"
                },
                "operator_username": {
                    "type": "string"
                },
                "safety_stock": {
                    "type": "integer"
                },
                "spec": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "internal_product.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code 是错误码。",
                    "type": "string"
                },
                "message": {
                    "description": "Message 是前端可展示错误信息。",
                    "type": "string"
                },
                "request_id": {
                    "description": "RequestID 是请求 ID，用于关联日志。",
                    "type": "string"
                }
            }
        },
        "internal_product.createProductRequest": {
            "type": "object",
            "required": [
                "code",
                "name",
                "operator_employee_id"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "example": "P-001"
                },
                "default_cost": {
                    "type": "integer",
                    "example": 10000
                },
                "name": {
                    "type": "string",
                    "example": "白色外壳"
                },
                "operator_employee_id": {
                    "type": "integer",
                    "example": 1
                },
                "safety_stock": {
                    "type": "integer",
                    "example": 10
                },
                "spec": {
                    "type": "string",
                    "example": "标准"
                },
                "unit": {
                    "type": "string",
                    "example": "个"
                }
            }
        },
...
        "internal_workorder.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code 是错误码。",
                    "type": "string"
                },
                "message": {
                    "description": "Message 是前端可展示错误信息。",
                    "type": "string"
                },
                "request_id": {
                    "description": "RequestID 是请求 ID，用于关联日志。",
                    "type": "string"
                }
            }
        },
...
        "internal_workorder.temporaryProductRequest": {
            "type": "object",
            "required": [
                "code",
                "name",
                "operator_employee_id"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "example": "P-001"
                },
                "name": {
                    "type": "string",
                    "example": "白色外壳"
                },
                "operator_employee_id": {
                    "type": "integer"
                },
                "spec": {
                    "type": "string",
                    "example": "标准"
                },
                "unit": {
                    "type": "string",
                    "example": "个"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT 登录凭证，格式为：Bearer \u003ctoken\u003e。",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}